splice functions from member.cpp
