+++ Address.h
@@ -19,29 +19,12 @@
 class Address {
 public:
     char addr[6];
-    Address() {}
-    // Copy constructor
+    Address();
     Address(const Address &anotherAddress);
-    // Overloaded = operator
     Address& operator =(const Address &anotherAddress);
     bool operator ==(const Address &anotherAddress);
-    Address(string address) {
-        size_t pos = address.find(":");
-        int id = stoi(address.substr(0, pos));
-        short port = (short)stoi(address.substr(pos + 1, address.size()-pos-1));
-        memcpy(&addr[0], &id, sizeof(int));
-        memcpy(&addr[4], &port, sizeof(short));
-    }
-    string getAddress() {
-        int id = 0;
-        short port;
-        memcpy(&id, &addr[0], sizeof(int));
-        memcpy(&port, &addr[4], sizeof(short));
-        return to_string(id) + ":" + to_string(port);
-    }
-    void init() {
-        memset(&addr, 0, sizeof(addr));
-    }
+    Address(string address);
+    string getAddress();
 };
 
 #endif /* Address_hpp */
